requirements.txt
runpod
torch
transformers
Pillow
accelerate
numpy
huggingface-hub
